+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+## [0.1.1] - 2026-01-03
+
+### Added
+- Allow skipping tag creation in `pps release` if the tag already exists.
+- Improved error handling for existing git tags.
+
 ## [0.1.0] - 2025-12-30
 
 ### Added